+++ Content/MusicBoxs/MusicEvent/MainThemeMagnoliaMusicBox.cs
@@ -1,11 +1,6 @@
 ﻿using EnderLiliesMusicPack.Common;
 using EnderLiliesMusicPack.Content.Rarity;
 using EnderLiliesMusicPack.Content.Tiles.MusicEvent;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using Terraria.ModLoader;
 
 namespace EnderLiliesMusicPack.Content.MusicBoxs.MusicEvent
@@ -14,6 +9,6 @@ namespace EnderLiliesMusicPack.Content.MusicBoxs.MusicEvent
     {
         public override string MusicName => MusicPathing.MainThemeMagnolia;
         public override int MusicBoxTile => ModContent.TileType<MainThemeMagnoliaMusicBoxTile>();
-        public override int LiliesOrLilac => LiliesRarityID.lilacID;
+        public override int LiliesOrLilac => (int)LiliesRarityID.lilacID;
     }
 }
